+++ packages/language-css/src/expressions/__tests__/performance/query-selectors.test.ts
@@ -43,6 +43,7 @@ suite("CSS - Query Selector Expression Factory", function() {
 
   suiteSetup(function() {
     expressions = querySelectorExpressionFactory({
+      kind: "class",
       prefix: "\\.",
     });
   });