+++ packages/language-js/src/__tests__/integration/class.test.ts
@@ -44,7 +44,9 @@ suite("The @webmangler/language-js plugin", function() {
     });
 
     test("has support for mangling 'query-selectors'", function() {
-      const options: QuerySelectorOptions = { };
+      const options: QuerySelectorOptions = {
+        kind: "class",
+      };
 
       const result = plugin.getExpressions("query-selectors", options);
       expect(result).to.have.length.above(0);