+++ packages/types/CHANGELOG.md
@@ -8,7 +8,8 @@ Versioning].
 
 ## [Unreleased]
 
-- _No changes yet_
+- Add case sensitivity flag to `CssDeclarationValueOptions` and
+  `QuerySelectorOptions`. ([#338])
 
 ## [0.1.26] - 2022-07-02
 
@@ -37,5 +38,6 @@ Versioning].
 [#306]: https://github.com/ericcornelissen/webmangler/pull/306
 [#311]: https://github.com/ericcornelissen/webmangler/pull/311
 [#314]: https://github.com/ericcornelissen/webmangler/pull/314
+[#338]: https://github.com/ericcornelissen/webmangler/pull/338
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/ "Keep a CHANGELOG"
 [semantic versioning]: https://semver.org/spec/v2.0.0.html "Semantic versioning"